requirements.txt
pygame>=2.5.0
opencv-python>=4.8.0
numpy>=1.24.0
pynput>=1.7.6
PyOpenGL>=3.1.5
imgui[pygame]>=2.0.0
zeroconf>=0.130.0
reedsolo>=1.7.0
av>=12.0.0
